+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## API
+The API consists of all public Java types from `com.atlassian.performance.tools.jiraperformancetests.api` and its subpackages:
+
+  * [source compatibility]
+  * [binary compatibility]
+  * [behavioral compatibility] with behavioral contracts expressed via Javadoc
+
+[source compatibility]: http://cr.openjdk.java.net/~darcy/OpenJdkDevGuide/OpenJdkDevelopersGuide.v0.777.html#source_compatibility
+[binary compatibility]: http://cr.openjdk.java.net/~darcy/OpenJdkDevGuide/OpenJdkDevelopersGuide.v0.777.html#binary_compatibility
+[behavioral compatibility]: http://cr.openjdk.java.net/~darcy/OpenJdkDevGuide/OpenJdkDevelopersGuide.v0.777.html#behavioral_compatibility
+
+
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/jira-performance-tests/branches/compare/master%0Drelease-0.1.1
 
@@ -12,6 +24,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Allow throttling virtual user diagnostics. 
+
+### Changed
+- Define the public API.
 
 ## [0.1.0]
 [0.1.0]: https://bitbucket.org/atlassian/jira-performance-tests/branches/compare/release-0.1.0%0Drelease-0.0.3